:recycle: [functional] Move fixtures from test to conftest.py
